Add documentation, fix broken tests, move default vocab tests to inflector tests.
