+++ src/Humanizer.Tests/ApiApprover/PublicApiApprovalTest.approve_public_api.approved.txt
@@ -243,6 +243,23 @@ public class In
     public System.DateTime TheYear(int year) { }
 }
 
+public class Vocabularies
+{
+    public Vocabularies() { }
+    public Humanizer.Inflections.Vocabulary Default { get; }
+}
+
+public class Vocabulary
+{
+    public Vocabulary() { }
+    public void AddIrregular(string singular, string plural) { }
+    public void AddPlural(string rule, string replacement) { }
+    public void AddSingular(string rule, string replacement) { }
+    public void AddUncountable(string word) { }
+    public string Pluralize(string word, bool inputIsKnownToBeSingular) { }
+    public string Singularize(string word, bool inputIsKnownToBePlural) { }
+}
+
 public class InflectorExtensions
 {
     public string Camelize(string input) { }
